pyproject.toml
[tool.poetry]
name = "arcipelago"
version = "0.1.0"
description = "A moderated event platform for Telegram."
authors = ["Flavio <flavio.petruzzellis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.14"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pre-commit = "^3.3.1"

[tool.poetry.group.gcalendar]
optional = true

[tool.poetry.group.gcalendar.dependencies]
google = "^3.0.0"
google-api-python-client = "^2.78.0"

[tool.pytest.ini_options]
pythonpath = "arcipelago"

[tool.poetry.scripts]
"arcipelago:run" = 'arcipelago.main:main'
"arcipelago:init-db" = 'arcipelago.init_db:main'
"arcipelago:migrate-db" = 'arcipelago.migrate:main'
"arcipelago:import-db" = 'arcipelago.db:import_db'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
